pyproject.toml
[project]
name = "demon-receipt-box"
version = "0.1.0"
description = "Voice recorder with local offline speech-to-text transcription"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "sounddevice",
    "scipy",
    "faster-whisper",
    "llama-cpp-python",
    "huggingface-hub",
    "pywin32; sys_platform == 'win32'",
    "Pillow",
]
